Сохранение в locaStorage, table в отдельном компоненте
